requirements.txt
Django>=3.0,<4.0
beautifulsoup4>=4.10.0 
djangorestframework>=3.13.1
python-dotenv>=0.20.0
drf-spectacular>=0.22.0  